Split math_utils from utils
